backend/requirements.txt
Click
Flask
Flask-Cors
Flask-Login
Flask-Bcrypt
flask_wtf
itsdangerous
Jinja2
MarkupSafe
six
Werkzeug
pymongo
pyyaml
